Editor Playlist Buttons Updated!
The play button functionality works from the playlist page and search result page. Administrators and Editors can now play the videos from their list(s) or the search result page.
The "Add New Song" is now moved to the top of the list under the title of the list from playlist index page.
The cursors are modified for the search buttons from Add New Songs page.